+++ src/services/colors/index.js
@@ -1,0 +1,2 @@
+export { isColorDark } from './is_color_dark';
+export { VISUALIZATION_COLORS } from './visualization_colors';